Try to be smart about choosing initial width of sidebar.
Try to avoid the sidebar covering the main content of the page when
opened by choosing an initial width such that the sidebar's right edge
is aligned with the dominant right edge of content on the page,
constrained to the range (200, 428) which is the minimum allowed size
for the sidebar and the previous default width respectively.

The dominant right edge is calculated by finding the the most frequent
right edge for paragraphs in the page, rounded to the nearest 20px
"stop", where paragraphs are weighted by the length of their text.

If no such edge can be found, the calculation bails out at the default
width is used.