Use pre-transpiled version of dom-anchor-fragment
dom-anchor-fragment's package.json specifies that the untranspiled
source should be used in the browser, together with Browserify
configuration.

This isn't compatible with the Babelify configuration that is used to
transpile the main client's code.

Resolve the issue by configuring the client explicitly to use the
pre-transpiled version of the source in the package.