Rename config() to configFrom()
Rename the `config()` function that creates the `options` object to
`configFrom()`.

This is because I want to rename the `options` object to `config` in a
future commit, so I need the `config()` function to be called something
else.

I think `configFrom(window)` is also clearer (that this is a function
that returns a `config` object created from the given window) than
`config(window)` (which could be read as configuring the window, for
example).