Split extractAnnotationQuery() into two functions
For a future refactoring of this config code I need to have one function
per config setting (the function returns the value of the config
setting).

extractAnnotationQuery() is actually a single function for two config
settings - it returns either {query: *} or {annotations: *} depending on
what's in the URL #fragment.

This doesn't fit into my planned refactoring, so split
extractAnnotationQuery() into two settings functions: one for
config.query and one for config.annotations.