Simplify reading query and annotations
There was some code complexity here aimed at ensuring that _either_
config.annotations or config.query would be set but never both. It isn't
really necessary for the code to maintain this condition, especially
given that the way that both settings are read from the URL fragment
means that both cannot be present at once anyway.