Save document metadata in app state.
In preparation for conditionally including 'doi:' URIs in the set of
search URIs depending on the active feature flags, store the complete
document metadata for connected frames in the app state, instead of just
extracting the document fingerprint and search URIs.

This will allow the code that searches for annotations to later extract
the set of search URIs just before it sends the query, at a point when
the client has feature flag information.