Move search URI extraction to a selector
Move the extraction of search URIs from frame metadata out of the
callback for the "getDocumentInfo" cross-frame request and into a
selector function.

This will make it easier to change the set of search URIs that the
client sends for a given set of connected frames depending on enabled
features. It also makes it easier to implement tests for this.

 * Remove `searchUris` from frame objects in the app state and instead
   compute this list on-demand from the frame URI and metadata when
   needed.