Access element as self.element
Adder() contains a `var self = this` line and a line that assigns the
`element` that `createAdderDOM()` returns to `this.element`, but it
usually accesses `element` via the local variable element (and
occassionally as `this.element`).

Change this to be consistent: assign `self.element = createAdderDOM()`;
always access it as `self.element` not `this.element`; delete the local
variable `element`.

This just makes the code a little simpler and more self-consistent.