Don't throw if hypothesisConfig() isn't a function
Log a warning, rather than throwing an error, if
window.hypothesisConfig() exists but isn't a function.

"Warn but continue" is in line with our general approach to handling
invalid configuration in the client. For example if there's an invalid
js-hypothesis-config script in the page it logs a warning and continues.

For some required config settings, such as config.app, if the setting is
invalid then the client crashes out immediately because it can't
continue. But otherwise it should warn and continue.