Fix a couple of config tests
Because these tests use a fake "foo" setting I could, for example,
whitelist the real openLoginForm setting as readable from the host page
by browser extensions and the tests would still be passing although they
should be failing.

Fix the tests to use real setting names.

Of course now the list of setting names in the test needs to be updated
whenever a setting is added, but that will be fixed later by better code
factoring.