Coordinate token refreshes between tabs.
Prevent multiple client instances in the same browser trying to refresh
OAuth tokens concurrently using the mutex service.

As part of this, the behavior of `tokenGetter` during a token refresh
was changed to wait until the refresh completes and then return the
new token, rather than returning the old token during the refresh.

The reason for this that it makes testing simpler when additional
async steps are added into the refresh process. Test code can simply
call `tokenGetter` after a token expires and await the result before
checking that the expected refresh steps happened.