Add a mutex service to help coordinate tasks between client instances
This service provides an API that can be used to prevent multiple
clients attempting to perform tasks (eg. refreshing OAuth tokens)
concurrently.

The implementation is currently based on localStorage and a simple spin
unlock mechanism. For a more sophisticated implementation, we may want
to use one of the existing npm packages for this.