Add tests for "since" and "uri" filters.
Codecov tells me that we are missing tests for these fields.

In the case of the "since" filter, this found an issue where `new
Date(<ann.updated value>)` was failing in PhantomJS because value
normalization lower-cased the 'T' and 'Z' chars in the ISO date string
and this caused parsing to fail. Fix this by making the `value` function
return a `Date` object which won't get normalized.