Remove "result" field handling from viewFilter service
This was undocumented and untested, and it is very unlikely that anyone
actually uses it.

It has also fundamentally been broken for a long time in the client
because filtering was applied _before_ sorting when building the thread
structure, so it was arbitrary (from the user's point of view) which
annotations would be kept by a `result` limit.