Remove the view switcher component
Remove the (feature-flagged) view switcher component.

We missed the chance to strike while the iron was hot on this one -
neither Dawa or I have time anymore to work on further design
iterations. I don't want to have dead code living behind a feature flag
so I'm removing it. The code will still be in git if we ever want to
revive it. Perhaps these tabs can be re-designed as part of a larger
re-design of the sidebar in the future.

I'll leave the GitHub issue about the usability issues with the existing
selection tabs open: https://github.com/hypothesis/product-backlog/issues/327

For the record the main remaining issues with this were:

1. An undesirable visual popping happens when the view switcher loads on
pages that have a lot of annotations. This also happens with the
existing selection tabs but is arguably more visible with the view
switcher.

2. When there are only two tabs (no orphans) the visual design of the
tabs doesn't make it as immediately obvious which is the currently
selected tab. The other tab may be mistaken for the selected one.

The TODO list was here:

https://github.com/hypothesis/product-backlog/issues/327#issuecomment-311073581

The pull requests for this view switcher were:

https://github.com/hypothesis/client/pull/429
https://github.com/hypothesis/client/pull/481
https://github.com/hypothesis/client/pull/482
https://github.com/hypothesis/product-backlog/issues/327