Require iframes to opt into annotation support
The client's iframe support is not yet robust enough (see
https://github.com/hypothesis/client/issues/530) to enable it
automatically for all iframes on arbitrary web pages.

To support the needs of EPUB viewers and others in the meantime while
preventing problems on eg. pages with larger numbers of iframed ads,
require the publisher to opt iframes into annotation by adding the
"enable-annotation" attribute to them.