Do not update tags in UPDATE_ANCHOR_STATUS action
Historically annotations did not have local tags until after they had
been anchored in the page.

This is no longer the case as annotations are assigned local tags as
soon as they are added into the app state by ADD_ANNOTATIONS.

Therefore, it is no longer necessary for the UPDATE_ANCHOR_STATUS action
to support updating the tags of annotations or to support identifying
the annotation which has been anchored by its ID instead of its tag.