Re-arrange OAuth service tests
Move the tests for fetching and persistence of tokens from/to storage
into the `describe('#tokenGetter', ...` block since this is the method
called from outside the service that initiates the fetching /
persistence.

Also reword a couple of the test cases to be clearer now that the OAuth
service now handles fetching tokens using methods other than exchanging
grant tokens.