Show display names for third-party users
Whenever possible, always show display names instead of usernames on
annotation cards if the creator of the card is a third-party user.

Even if the client_display_names feature flag is turned off, display
names _will_ still be shown for third-party users.

Display names will _not_ be shown for first-party users, unless the
client_display_names feature flag is on.

If the api_render_user_info feature flag is turned off in h, or if the
user doesn't have a display name, then of course the username must be
shown as the client has no access to a display name.