Extract logic for interacting with OAuth server into a separate module
In order to make the `oauth-auth` service easier to follow and test,
split out the logic responsible for actually making requests to the
OAuth endpoints into a separate module.

The `oauth-client` module is also intended to be something that a third
party could fairly easily reuse or study for building their own client.