Change button tag to span for the minimal onboarding look.
The new note button in the onboarding tutorial shouldn't be a button because it isn't clickable.
It is there to teach the user what to look for to create a new note.
Turn the button element into a span with the same styling as the
new note button which appears in the 'Page Notes' tab.

See hypothesis/product-backlog#351
and https://github.com/hypothesis/client/pull/580#discussion_r151372544