Don't link to annotation pages that don't exist
If an annotation, as received from the API, doesn't have any html link
in the links object in the API response, then `vm.href` in
`timestamp.html` will be an empty string (due to the behaviour of the
`htmlLink` function in `annotation-header.js`).

When this happens render the annotation's timestamp as a `<span>` rather
than as an `<a>` with an `href` attribute that has no value.