Don't show page share button on eLife pages
Don't show the "Share this page" button in the top-right of the sidebar
on pages, such as eLife's, that have both an embedded Hypothesis client
and a services config including their own `authority` string.

The "Share this page" button is a Via link, and Via doesn't work with
pages using a third-party authority config.

Fixes https://github.com/hypothesis/client/issues/615