Move annotationComponent() into scope
Move the annotationComponent() function into the describe() scope nearer
to where it's actually used. This puts the code nearer to where it's
used and is going to be useful in the next commit when it's going to be
changed to make use of one of the fake variables in that scope.