Update an out-of-date fake
The fakeAccountID object in annotation-test.js was out of date: it
contained only a username() function that annotation.js doesn't call
anymore, and did not contain the isThirdPartyUser() function that
annotation.js _does_ call.

The tests were still passing because annotation.js's isThirdPartyUser()
method, which call's account-id's isThirdPartyUser(), wasn't tested. Add
an integration test for this method to catch this.

I've also moved fakeAccountID out into its own variable with a name,
instead of it just being an anonymous object literal. This makes it more
explicit. For example if account-id's API is changed, you can just grep
the code for fakeAccountID to find all the fakes that need to be
updated.