Add new config option that scopes annotation sync to just sub frames
This config is helpful when using multi-frame support. In some cases, annotating content should only happen at the sub frame level because if similar content is being annotated at either the sub frame level or the top frame level, the anchoring of annotations may not work as intended. Please see #631 for more details.