Get the list of groups from the new endpoint and update it wherever applicable.
Fetch the list of groups from the new `/api/groups` endpoint instead of
using the groups list returned in the `/api/profile` result. This
supports providing the current document URI as a parameter, which
enables the server to change the list of open groups returned depending
on the current URL.

In order to update the groups list correctly, taking into account the
current document URI, when the user joins or leaves a group, the
`streamer` service now handles "session change" notifications by
triggering a re-fetch of groups from the `/api/groups` endpoint instead
of updating the groups directly from the payload of the WebSocket
message.

Co-authored-by: Robert Knight <robertknight@gmail.com>