Add store module for groups
Add a Redux module that stores state related to groups, including the
currently focused group and the list of all known groups.

This is an initial step towards moving this state from the groups
service into the store, which will make reacting to changes in the set
of loaded groups, or the focused group, easier.