Remove activity page links for third-party open groups
Activity pages are currently only supported for open groups belonging to
the h service's first party authority.

The team has agreed that the best way to resolve this would be to only
include the "activity page" link in the `/api/groups` response for
groups where that is supported. However to give us more time to hammer
out the API details, we are committing a small change to the client in
the interim.