Update frame metadata in sidebar when notification is received
Implement an `UPDATE_FRAME` action which updates the URI and metadata
for an existing frame and call it when an "updateFrame" message is
received from the annotator in the host page. This in turn will trigger
the client to unload any annotations associated with the frame's
previous URL and load annotations for the new URL.

Together with the previous commits this enables the client to respond to
URL changes in single page applications and sites that use PJAX to
optimize route changes, such as GitHub and Medium.