Rename src/sidebar/{store => services/api}
Give the service which provides a client for the Hypothesis API a more
obvious name and place it in a `services/` dir in accordance with the
plan outlined in #687. The previous name of "store" was particularly
confusing because of the existence of another kind of unrelated store
(the Redux one) in the application.

The service was previously registered with the `factory` method but has
the same structure as all the other services which are registered with
`service`, so I changed that for consistency.