Move services to `services/` directory
Following the plan outlined in #687, move services into a `services/`
directory.

The definition of "services" in that plan (a possibly stateful module
that handles interaction with an external party) is slightly more narrow
than what constitutes an Angular service (a stateful singleton that has
dependencies on injectable services/values).

However as a first approximation most Angular services are services as
defined in #687, so I've simply moved them all to the `services/`
directory.