Extract `createStore` helper out of `store` service
The `store` factory function contains some useful logic for composing a
store out of a set of _modules_ which handle different parts of the
application state, exposing the actions and selectors from each module
as methods on the store.

To facilitate re-use of this logic in tests and to make it easier to
test this logic directly, extract it out of the `store` function into a
separate `createStore` helper.

Doing this enables removal of `modules/index.js` which existed to
combine parts of the modules together.