Always include `thunk` middleware in stores created by `createStore`
The `thunk` middleware is used by the `annotations` module and likely
others in future. Rather than require it to be passed to `createStore`
in each test that uses it, I think it makes sense to always include it
as the standard way to handle async/non-pure actions in the client.