Always get metadata from injected DOM Document
The `Document` class for reading DOM document metadata takes an optional
DOM `Document` constructor argument as a test seam.

However the actual code only used the injected document in some places,
in others it used jQuery which in turn reads from the global
`window.document`.

Refactor this code to use the injected DOM Document everywhere and
remove the jQuery dependency.