Fix test failure in Chrome caused by lack of isolation between tests
A `<link>` tag with a relative URL from a previous test was causing
problems in Chrome when testing the behavior of `Document#uri` in
documents with a non-HTTP/file `<base>`.

That is an edge case which we may want to cover at some point, but the
more significant issue is that the `Document#uri` tests were not
isolated from the earlier ones. Modify the `createDoc` helper so that it
uses a blank HTML document instead of using the global `document`
object.