Add documentation for externalContainerSelector
Document externalContainerSelector configuration option.

This is marked as experimental since it is new and may change based on
feedback from early users (eg. Atypon).

This API is also currently incomplete in that there are a number of
additional APIs needed to fully implement a custom sidebar container
which do not exist, eg. the ability for the host page to know when to
show or hide the sidebar as the user interacts with highlights in the
page.